+++ src/infrastructure/dto/product_dto.go
@@ -3,5 +3,5 @@ package dto
 type ProductDTO struct{
 	ID 	  uint
 	Name  string
-	Price uint64
+	Price float64
 }
